+++ yo/concrete/bisonc++/parser/parser.h
@@ -70,7 +70,7 @@ class Parser: public ParserBase
     // support functions for parse():
 
         void executeAction(int d_production);
-        unsigned errorRecovery();
+        size_t errorRecovery();
         int lookup(int token);
         int nextToken();
 };
