+++ yo/concrete/bisonc++/parser/parserbase.h
@@ -33,7 +33,7 @@ struct STYPE
 
     private:
         int d_stackIdx;
-        std::vector<unsigned>   d_stateStack;
+        std::vector<size_t>   d_stateStack;
         std::vector<STYPE>      d_valueStack;
 
     protected:
@@ -48,9 +48,9 @@ struct STYPE
             UNEXPECTED_TOKEN,
         };
         bool        d_debug;
-        unsigned    d_nErrors;
+        size_t    d_nErrors;
         int         d_token;
-        unsigned    d_state;
+        size_t    d_state;
         STYPE      *d_vsp;
         STYPE       d_val;
 
@@ -65,14 +65,14 @@ struct STYPE
         {
             return d_debug;
         }
-        void pop(unsigned count = 1);
-        void push(unsigned nextState);
-        unsigned reduce(PI const &productionInfo);
+        void pop(size_t count = 1);
+        void push(size_t nextState);
+        size_t reduce(PI const &productionInfo);
         void setDebug(bool mode)
         {
             d_debug = mode;
         }
-        unsigned top() const;
+        size_t top() const;
 
 // class ParserBase ends
 };
