+++ yo/concrete/examples/pipe.h
@@ -10,7 +10,7 @@
             void readFrom(int fileDescriptor);
             int writeOnly();
             void writtenBy(int fileDescriptor);
-            void writtenBy(int const *fileDescriptors, unsigned n = 2);
+            void writtenBy(int const *fileDescriptors, size_t n = 2);
         private:
             void redirect(int d_fd, int alternateFd);
     };
