+++ yo/concrete/fork2/opipe/opipe.cc
@@ -1,6 +1,6 @@
 #include "opipe.ih"
 
-OPipe::OPipe(unsigned size)
+OPipe::OPipe(size_t size)
 :
     Pipe(),
     OFdNStreambuf(Pipe::d_fd[READ], size == 0 ? 1 : size),
