+++ yo/concrete/fork2/pipe/writevia2.cc
@@ -1,10 +1,10 @@
 #include "pipe.ih"
 
-void Pipe::writeVia(int const *fd, unsigned n)
+void Pipe::writeVia(int const *fd, size_t n)
 {
     close(d_fd[READ]);
 
-    for (unsigned idx = 0; idx < n; idx++)
+    for (size_t idx = 0; idx < n; idx++)
         dup(WRITE, fd[idx]);
 
     close(d_fd[WRITE]);
