+++ yo/stl/examples/not.cc
@@ -7,15 +7,15 @@
     int main(int argc, char **argv)
     {
         bool bArr[] = {true, true, true, false, false, false};
-        unsigned const bArrSize = sizeof(bArr) / sizeof(bool);
+        size_t const bArrSize = sizeof(bArr) / sizeof(bool);
 
-        for (unsigned idx = 0; idx < bArrSize; ++idx)
+        for (size_t idx = 0; idx < bArrSize; ++idx)
             cout << bArr[idx] << " ";
         cout << endl;
 
         transform(bArr, bArr + bArrSize, bArr, logical_not<bool>());
 
-        for (unsigned idx = 0; idx < bArrSize; ++idx)
+        for (size_t idx = 0; idx < bArrSize; ++idx)
             cout << bArr[idx] << " ";
         cout << endl;
     }
